fix: refactored deleteUser endpoint

+++ src/controllers/userController/deleteUser.ts
@@ -1,12 +1,23 @@
-import { Response, NextFunction, Request } from 'express';
+import { Response, NextFunction } from 'express';
 import { getManager } from 'typeorm';
 
+import { tryCatch } from '@src/utils/tryCatch';
+import { AppError } from '@src/utils/appError';
+import { RequestUser } from '@src/types/index';
 import { User } from '@src/models/entities/User';
 
-export const deleteUser = async (request: Request, response: Response, next: NextFunction) => {
+export const deleteUser = tryCatch(async (request: RequestUser, response: Response, next: NextFunction) => {
   const entityManager = getManager();
   const { id } = request.body;
-  // TODO add validation
+
+  if (!id) {
+    return next(new AppError('Please provide user ID', 400));
+  }
+
+  if (typeof id !== 'number') {
+    return next(new AppError('User ID should be a number', 400));
+  }
+
   const [user] = await User.findByIds([id]);
 
   await entityManager.delete(User, id);
@@ -15,4 +26,4 @@ export const deleteUser = async (request: Request, response: Response, next: Nex
     status: 'success',
     data: undefined,
   });
-};
+});
